pyproject.toml
[tool.poetry]
name = "liberte"
version = "0.1.0"
description = "Meme Coin Trading Bot"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.12"
aiokafka = "0.10.0"
aiohttp = "3.9.4"
websockets = "12.0"
prometheus-client = "0.14.1"
pyyaml = "6.0.1"
structlog = "24.1.0"
hvac = "2.3.0"
kafka-python = "^2.2.15"
confluent-kafka = "^2.11.1"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
flake8 = "^7.0"
cython = "^3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
collector = "collector.collector:main"

[[tool.poetry.packages]]
include = "collector"
from = "src"